+++ docs/assets/components-DwPP67yX.js
@@ -1,1 +1,0 @@
-import"./iframe-C1D-rtQk.js";import"./chunk-W22LQPXL-DCxhZZZG.js";import"./react-dom-DHlK3uMh.js";import{at as e,ot as t}from"./DocsRenderer-JROSPFPF-C4eVPPPQ.js";t();export{e as createCopyToClipboardFunction};