+++ src/components/ui/DataTable/CellEditor.tsx
@@ -1,5 +1,6 @@
 import { type KeyboardEvent } from 'react'
-import { cn, getEditClass, getEditAccent } from './helpers'
+import { cn } from '../../../lib/cn'
+import { getEditClass, getEditAccent } from './helpers'
 import type { CellValue, ColumnEditor } from './types'
 import type { ColorScheme } from '../../../types'